LayoutTests/platform/ios-simulator/fast/dom/Element/scrollLeft-Quirks-expected.txt

Tests that for quirks mode document.body.scrollLeft returns the scroll left value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL window.pageXOffset should be 500. Was 0.
FAIL document.body.scrollLeft should be 500. Was 0.
PASS document.documentElement.scrollLeft is 0
PASS successfullyParsed is true

TEST COMPLETE
